+++ fili-core/src/main/java/com/yahoo/bard/webservice/data/time/GranularityDictionary.java
@@ -1,0 +1,13 @@
+// Copyright 2016 Yahoo Inc.
+// Licensed under the terms of the Apache license. Please see LICENSE.md file distributed with this work for terms.
+package com.yahoo.bard.webservice.data.time;
+
+import com.yahoo.bard.webservice.druid.model.query.Granularity;
+
+import java.util.HashMap;
+
+/**
+ * Map of Granularity name to Granularity.
+ */
+public class GranularityDictionary extends HashMap<String, Granularity> {
+}