+++ CHANGELOG.md
@@ -24,6 +24,9 @@ Current
 
 ### Changed:
 
+- [The `getGrainMap` method in `StandardGranularityParser` class is renamed to `getDefaultGrainMap` and is made public static.](https://github.com/yahoo/fili/pull/88)
+    - Created new class `GranularityDictionary` and bind getGranularityDictionary to it
+
 - [`PhysicalTable` now uses `getAvailableIntervals` internally rather than directly referencing its intervals](https://github.com/yahoo/fili/pull/79)
 
 - [CSV attachment name for multi-interval request now contain '__' instead of ','](https://github.com/yahoo/fili/pull/76)